refactor: extract commitWorkspaceState into shared cli/src/lib/workspace-git.ts
